JobCalculate

+++ Exposure/Exposure.Web/Controllers/JobsController.cs
@@ -24,19 +24,15 @@ namespace Exposure.Web.Controllers
 
             ViewBag.Skills = new SelectList(db.Skills, "SkillID", "SkillDescription");
             var jobs = db.Jobs.Include(j => j.Employer);
-            //var draw = Request.Form.GetValues("draw").FirstOrDefault();
-            //var start = Request.Form.GetValues("start").FirstOrDefault();
-            //var length = Request.Form.GetValues("length").FirstOrDefault();
+            var jobHistory = db.JobApplications.Include(j => j.Job.Employer);
 
-            //int pageSize = length != null ? Convert.ToInt32(length) : 0;
-            //int skip = start != null ? Convert.ToInt32(start) : 0;
-            //int recordsTotal = 0;
+
 
             if (User.IsInRole("Admin"))
             {
                 jobs = db.Jobs.Include(j => j.Employer).Include(j => j.Skill).Include(j => j.Suburb).OrderBy(j => j.DateAdvertised);
-            }            
-            else if(User.IsInRole("Employer"))
+            }
+            else if (User.IsInRole("Employer"))
             {
                 var userId = User.Identity.GetUserId();
                 jobs = db.Jobs.Include(j => j.Employer).Include(j => j.Skill).Include(j => j.Suburb).OrderBy(j => j.DateAdvertised).Where(j => j.Employer.EmployerID == userId);
@@ -44,7 +40,7 @@ namespace Exposure.Web.Controllers
             }
             else if (User.IsInRole("Worker"))
             {
-                var jobHistory = db.JobApplications.Include(j => j.Job.Employer).Include(j => j.Job).Include(j => j.WorkerID.Equals(id));
+                jobHistory = db.JobApplications.Include(j => j.Job.Employer).Include(j => j.Job).Include(j => j.WorkerID.Equals(id)).OrderByDescending(x => x.Job.StartDate).OrderBy(x => x.Job.Completed);
             }
 
 
@@ -85,16 +81,17 @@ namespace Exposure.Web.Controllers
             //ViewBag.JobID = job;
             //ViewBag.JList = jobs;
             ViewBag.WList = workers;
-            //ViewBag.Jobs = jobs;
+            ViewBag.Jobs = jobs;
+            ViewBag.JobHistory = jobHistory;
 
             return View(jobs);
-                                 
-            
+
+
         }
 
         [AllowAnonymous]
-        public ActionResult Search(string currentFilter,string sortOrder, int? skill, string search, int? page, int? location)
-        {                      
+        public ActionResult Search(string currentFilter, string sortOrder, int? skill, string search, int? page, int? location)
+        {
 
             if (search != null)
             {
@@ -105,13 +102,13 @@ namespace Exposure.Web.Controllers
                 search = currentFilter;
             }
 
-            ViewBag.CurrentFilter = search;              
+            ViewBag.CurrentFilter = search;
 
-            ViewBag.skill = new SelectList(db.Skills.OrderBy(x=>x.SkillDescription), "SkillID", "SkillDescription");
+            ViewBag.skill = new SelectList(db.Skills.OrderBy(x => x.SkillDescription), "SkillID", "SkillDescription");
             ViewBag.TitleSortParm = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
             ViewBag.DateSortParm = sortOrder == "Date" ? "date_desc" : "Date";
             ViewBag.location = new SelectList(db.Suburbs, "SuburbID", "SubName");
-            var jobs = db.Jobs.Include(j => j.Employer).Include(j => j.Skill).Include(j => j.Suburb).Where(c=>c.Completed==false);
+            var jobs = db.Jobs.Include(j => j.Employer).Include(j => j.Skill).Include(j => j.Suburb).Where(c => c.Completed == false);
 
             switch (sortOrder)
             {
@@ -134,32 +131,32 @@ namespace Exposure.Web.Controllers
                     jobs = jobs.OrderBy(j => j.DateAdvertised);
                     break;
             }
-        
+
             if (skill != null)
-            {                              
-                   jobs = jobs.Where(j => j.SkillID==skill);             
-                                
+            {
+                jobs = jobs.Where(j => j.SkillID == skill);
+
             }
 
-            if (location!=null)
+            if (location != null)
             {
-                jobs = jobs.Where(j => j.Suburb.SuburbID==location);
+                jobs = jobs.Where(j => j.Suburb.SuburbID == location);
             }
 
             if (!String.IsNullOrEmpty(search))
             {
-                jobs = jobs.Where(j => j.Title.Contains(search));                                        
+                jobs = jobs.Where(j => j.Title.Contains(search));
             }
-            
+
             //int pageSize = 3;
             //int pageNumber = (page ?? 1);
             var rows = jobs.Count();
-            if(rows==0)
+            if (rows == 0)
             {
                 TempData["Empty"] = "No jobs available. Please edit your search criteria or check back later";
-                
+
             }
-            
+
             ViewBag.Jobs = jobs;
 
             return View(jobs);
@@ -168,7 +165,7 @@ namespace Exposure.Web.Controllers
         public ActionResult History(string SortOrder, string id, int? skill, DateTime? startDate, string search)
         {
             var userId = User.Identity.GetUserId();
-            var jobsHistory = db.JobApplications.Include(j => j.Job.Employer).Include(j => j.Job.Skill).Include(j => j.Job.Suburb).Where(j => j.Job.Employer.EmployerID.Equals(id)).OrderBy(j => j.Job.DateAdvertised).Include(w=>w.WorkerID.Equals(userId));
+            var jobsHistory = db.JobApplications.Include(j => j.Job.Employer).Include(j => j.Job.Skill).Include(j => j.Job.Suburb).Where(j => j.Job.Employer.EmployerID.Equals(id)).OrderBy(j => j.Job.DateAdvertised).Include(w => w.WorkerID.Equals(userId));
 
             if (!String.IsNullOrEmpty(id))
             {
@@ -197,7 +194,7 @@ namespace Exposure.Web.Controllers
 
         public ActionResult JobApplication(int? id)
         {
-            return RedirectToRoute("Default", new { controller = "JobApplications", action = "Create", id=id });
+            return RedirectToRoute("Default", new { controller = "JobApplications", action = "Create", id = id });
         }
 
         ////GET: Jobs/Update
@@ -219,21 +216,21 @@ namespace Exposure.Web.Controllers
             var AppCheck = db.JobApplications.Where(j => j.JobID == (id)).Where(j => j.Response == Reply.Hired);
             var rows = AppCheck.Count();
 
-            if(rows==0)
+            if (rows == 0)
             {
                 TempData["Application"] = "No application has been accepted/submitted for this job yet.";
             }
 
-            if(job.Completed==true)
+            if (job.Completed == true)
             {
                 TempData["Completed"] = "Job already completed";
             }
-            
+
             ViewBag.Job = db.Jobs.Where(j => j.JobID == id);
             ViewBag.JobID = id;
             ViewBag.JobStart = Convert.ToDateTime(job.StartDate);
             ViewBag.JobEnd = Convert.ToDateTime(job.EndDate);
-            ViewBag.SuburbID = new SelectList(db.Suburbs,"SuburbID", "SubName", job.SuburbID);
+            ViewBag.SuburbID = new SelectList(db.Suburbs, "SuburbID", "SubName", job.SuburbID);
             return View(job);
 
         }
@@ -241,12 +238,12 @@ namespace Exposure.Web.Controllers
         //POST: Jobs/Details
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles="Employer")]
+        [Authorize(Roles = "Employer")]
         public ActionResult Update([Bind(Include = "JobID,Completed")]Job job)
-        {           
+        {
             Job a = db.Jobs.Find(job.JobID);
 
-            a.Completed = job.Completed;     
+            a.Completed = job.Completed;
 
             var state = ModelState;
             try
@@ -261,8 +258,8 @@ namespace Exposure.Web.Controllers
                 ViewBag.JobID = job.JobID;
                 ViewBag.SuburbID = new SelectList(db.Suburbs, job.SuburbID);
                 return View(job);
-            }        
-                        
+            }
+
         }
 
         // GET: Jobs/Details/5
@@ -280,15 +277,33 @@ namespace Exposure.Web.Controllers
             return View(job);
         }
 
+        public JsonResult Calculate(int id, DateTime start, DateTime end, DateTime sTime, DateTime eTime)
+        {
+            Skill sk = db.Skills.Find(id);
+
+            var st = Convert.ToDouble(sTime.Hour);
+            var et = Convert.ToDouble(eTime.Hour);
+
+            var Days = end - start;
+
+            var hours = et - st;
+
+            var PayPerDay = sk.Recom_Rate * hours;
+            var skill = sk.SkillDescription;
+
+            return Json(PayPerDay, skill);
+
+        }
+
         // GET: Jobs/Create
-        [Authorize(Roles =("Admin, Employer"))]
+        [Authorize(Roles = ("Admin, Employer"))]
         public ActionResult Create()
         {
-            var skills = db.Skills.ToList().OrderBy(s=>s.SkillDescription);
+            var skills = db.Skills.ToList().OrderBy(s => s.SkillDescription);
             ViewBag.Skills = skills;
             ViewBag.EmployerName = User.Identity.Name;
-            ViewBag.SkillID = new SelectList(db.Skills.OrderBy(x=>x.SkillDescription), "SkillID", "SkillDescription");
-            ViewBag.SuburbID = new SelectList(db.Suburbs.OrderBy(s=>s.SubName), "SuburbID", "SubName");
+            ViewBag.SkillID = new SelectList(db.Skills.OrderBy(x => x.SkillDescription), "SkillID", "SkillDescription");
+            ViewBag.SuburbID = new SelectList(db.Suburbs.OrderBy(s => s.SubName), "SuburbID", "SubName");
             return View();
         }
 
@@ -297,94 +312,37 @@ namespace Exposure.Web.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles =("Admin, Employer"))]
-        public ActionResult Create([Bind(Exclude ="EmployerID, DateAdvertised", Include = "JobID,Title,StartDate, EndDate,SkillID,Description,StartTime,EndTime,Rate,SuburbID, AddressLine1,AddressLine2")] Job job)
+        [Authorize(Roles = ("Admin, Employer"))]
+        public ActionResult Create([Bind(Exclude = "EmployerID, DateAdvertised", Include = "JobID,Title,StartDate, EndDate,SkillID,Description,StartTime,EndTime,Rate,SuburbID, AddressLine1,AddressLine2")] Job job)
         {
             job.EmployerID = User.Identity.GetUserId();
-
+            //var sDate = job.StartDate;
             job.DateAdvertised = DateTime.UtcNow;
-
             var state = ModelState;
             if (ModelState.IsValid)
             {
                 db.Jobs.Add(job);
                 db.SaveChanges();
                 TempData["JobSuccess"] = "Job Successfully created";
-                return RedirectToRoute("Default", new { controller = "Jobs", action = "Index", id=User.Identity.GetUserId() });
+                return RedirectToRoute("Default", new { controller = "Jobs", action = "Index", id = User.Identity.GetUserId() });
             }
 
-            
+            var skills = db.Skills.ToList().OrderBy(s => s.SkillDescription);
             ViewBag.EmployerName = User.Identity.Name;
-            ViewBag.SkillID = new SelectList(db.Skills.OrderBy(x=>x.SkillDescription), "SkillID", "SkillDescription", job.SkillID);
-            ViewBag.SuburbID = new SelectList(db.Suburbs.OrderBy(s=>s.SubName), "SuburbID", "SubName", job.SuburbID);
+            ViewBag.Skills = skills;
+            ViewBag.SkillID = new SelectList(db.Skills.OrderBy(x => x.SkillDescription), "SkillID", "SkillDescription", job.SkillID);
+            ViewBag.SuburbID = new SelectList(db.Suburbs.OrderBy(s => s.SubName), "SuburbID", "SubName", job.SuburbID);
             TempData["JobSuccess"] = "Job could not be created. Please try again.";
             return View(job);
         }
 
         // GET: Jobs/Edit/5
         public ActionResult Edit(int? id)
-        {           
-                    
+        {
+
 
             ViewBag.jobSkills = new SelectList(db.Skills, "SkillID", "SkillDescription");
 
-            if(id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Job job = db.Jobs.Find(id);
-            if(job==null)
-            {
-                return HttpNotFound();
-            }
-
-            ViewBag.SuburbID = new SelectList(db.Suburbs, "SuburbID", "SubName", job.SuburbID);
-            ViewBag.JobID = id;
-
-            return View(job);
-            
-        }
-
-        // POST: Jobs/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Exclude ="DateAdvertised, Title",Include = "JobID,Title,DateAdvertised,SkillID,Description,StartTime,EndTime,Rate,SuburbID, AddressLine1,AddressLine2,StartDate, EndDate")] Job job)
-        {
-            Job j = db.Jobs.Find(job.JobID);
-                                
-            j.Description = job.Description;
-            j.StartTime = job.StartTime;
-            j.EndDate = job.EndDate;
-            j.EndTime = job.EndTime;
-            j.Rate = job.Rate;
-            j.SuburbID = job.SuburbID;
-            j.AddressLine1 = job.AddressLine1;
-            j.AddressLine2 = job.AddressLine2;
-
-           
-                try
-                {
-                    db.Entry(j).State = EntityState.Modified;
-                    db.SaveChanges();
-                    return RedirectToAction("Index", new {id=User.Identity.GetUserId() });
-                }
-                catch
-                {
-                    db.SaveChanges();
-                    return RedirectToAction("Index", new { id = User.Identity.GetUserId() });
-                }
-                    
-                        
-            ViewBag.jobSkills = new SelectList(db.Skills, "SkillID", "SkillDescription");
-            ViewBag.JobID = j.JobID;
-            return View(job);
-        }
-
-        // GET: Jobs/Delete/5
-        public ActionResult Delete(int? id)
-        {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -394,18 +352,82 @@ namespace Exposure.Web.Controllers
             {
                 return HttpNotFound();
             }
+
+            ViewBag.SuburbID = new SelectList(db.Suburbs, "SuburbID", "SubName", job.SuburbID);
+            ViewBag.JobID = id;
+
+            ViewBag.StartDate = job.StartDate;
+            ViewBag.EndDate = job.EndDate;
+
+
+
+            return View(job);
+
+        }
+
+        // POST: Jobs/Edit/5
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Edit([Bind(Exclude = "DateAdvertised, Title", Include = "JobID,Title,DateAdvertised,SkillID,Description,StartTime,EndTime,Rate,SuburbID, AddressLine1,AddressLine2,StartDate, EndDate")] Job job)
+        {
+            Job j = db.Jobs.Find(job.JobID);
+
+            j.Description = job.Description;
+            j.StartTime = job.StartTime;
+            j.StartDate = job.StartDate;
+            j.EndDate = job.EndDate;
+            j.EndTime = job.EndTime;
+            j.Rate = job.Rate;
+            j.SuburbID = job.SuburbID;
+            j.AddressLine1 = job.AddressLine1;
+            j.AddressLine2 = job.AddressLine2;
+
+            if(ModelState.IsValid)
+            {
+                db.Entry(j).State = EntityState.Modified;
+                db.SaveChanges();
+                return RedirectToAction("Index", new { id = User.Identity.GetUserId() });
+            }
+            
+
+            ViewBag.jobSkills = new SelectList(db.Skills, "SkillID", "SkillDescription");
+            ViewBag.JobID = j.JobID;
             return View(job);
         }
 
+        // GET: Jobs/Delete/5
+        //public ActionResult Delete(int? id)
+        //{
+        //    if (id == null)
+        //    {
+        //        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+        //    }
+        //    Job job = db.Jobs.Find(id);
+        //    if (job == null)
+        //    {
+        //        return HttpNotFound();
+        //    }
+        //    return View(job);
+        //}
+
         // POST: Jobs/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        public JsonResult Delete(int id)
         {
+            bool result = false;
+
             Job job = db.Jobs.Find(id);
-            db.Jobs.Remove(job);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (job != null)
+            {
+                db.Jobs.Remove(job);
+                db.SaveChanges();
+                result = true;
+            }
+
+            return Json(result, JsonRequestBehavior.AllowGet);
         }
 
         protected override void Dispose(bool disposing)
